requirements.txt
apache-airflow==2.8.0
kafka-python==2.0.2
requests==2.31.0
pandas==2.0.3
